Various tweaks
* Move the representation types into their own module so we can
  do the generic stuff in its own module and still use default
  signatures
