+++ src/Staged/GHC/Generics/Types.hs
@@ -1,4 +1,6 @@
 {-# LANGUAGE DataKinds            #-}
+{-# LANGUAGE DefaultSignatures    #-}
+{-# LANGUAGE FlexibleContexts     #-}
 {-# LANGUAGE PolyKinds            #-}
 {-# LANGUAGE ScopedTypeVariables  #-}
 {-# LANGUAGE StandaloneDeriving   #-}
@@ -37,6 +39,8 @@ module Staged.GHC.Generics.Types (
     -- * Generic type classes
     Generic (..),
     Generic1 (..),
+    genericTo,
+    genericFrom,
     -- * TH Types
     Code, Quote,
     -- * Utilities
@@ -44,93 +48,10 @@ module Staged.GHC.Generics.Types (
 ) where
 
 import Data.Kind           (Type)
-import GHC.Generics
-       (Associativity (..), C, Constructor (..), D, Datatype (..),
-       DecidedStrictness (..), Fixity (..), FixityI (..), Meta (..), R, S,
-       Selector (..), SourceStrictness (..), SourceUnpackedness (..))
-import GHC.TypeLits        (ErrorMessage (..), TypeError)
-import Language.Haskell.TH (Code, Quote)
+import Staged.GHC.Generics.RepTypes
+import Staged.GHC.Generics.GHCish (GGeneric (..), ghcGenericTo, ghcGenericFrom)
 
 import qualified GHC.Generics as GHC
-
--------------------------------------------------------------------------------
--- Combinators
--------------------------------------------------------------------------------
-
-data V2 (q :: z) (p :: k)
-
-deriving instance Eq (V2 q p)
-deriving instance Ord (V2 q p)
-deriving instance Show (V2 q p)
-deriving instance Read (V2 q p)
-
-data U2 (q :: z) (p :: k) = U2
-
-deriving instance Eq (U2 q p)
-deriving instance Ord (U2 q p)
-deriving instance Show (U2 q p)
-deriving instance Read (U2 q p)
-
-newtype M2 (i :: Type) (c :: Meta) (f :: z -> k -> Type) (q :: z) (p :: k)
-    = M2 { unM2 :: f q p }
-
-deriving instance Eq (f q p) => Eq (M2 i c f q p)
-deriving instance Ord (f q p) => Ord (M2 i c f q p)
-deriving instance Show (f q p) => Show (M2 i c f q p)
-deriving instance Read (f q p) => Read (M2 i c f q p)
-
-newtype K2 c (q :: Type -> Type) (p :: k)
-    = K2 { unK2 :: q c }
-
-deriving instance Eq (q c) => Eq (K2 c q p)
-deriving instance Ord (q c) => Ord (K2 c q p)
-deriving instance Show (q c) => Show (K2 c q p)
-deriving instance Read (q c) => Read (K2 c q p)
-
-infixr 5 :++:
-data ((f :: z -> k -> Type) :++: (g :: z -> k -> Type)) (q :: z) (p :: k)
-    = L2 (f q p)
-    | R2 (g q p)
-
-deriving instance (Eq (f q p), Eq (g q p)) => Eq ((f :++: g) q p)
-deriving instance (Ord (f q p), Ord (g q p)) => Ord ((f :++: g) q p)
-deriving instance (Show (f q p), Show (g q p)) => Show ((f :++: g) q p)
-deriving instance (Read (f q p), Read (g q p)) => Read ((f :++: g) q p)
-
-infixr 6 :**:
-data ((f :: z -> k -> Type) :**: (g :: z -> k -> Type)) (q :: z) (p :: k)
-    = f q p :**: g q p
-
-deriving instance (Eq (f q p), Eq (g q p)) => Eq ((f :**: g) q p)
-deriving instance (Ord (f q p), Ord (g q p)) => Ord ((f :**: g) q p)
-deriving instance (Show (f q p), Show (g q p)) => Show ((f :**: g) q p)
-deriving instance (Read (f q p), Read (g q p)) => Read ((f :**: g) q p)
-
--- https://gitlab.haskell.org/ghc/ghc/-/issues/15969#note_164233 !!!
-infixl 7 :@@:
-newtype ((f :: z -> k2 -> Type) :@@: (g :: k1 -> k2)) (q :: z) (p :: k1)
-    = App2 { unApp2 :: f q (g p) }
-
-deriving instance Eq (f q (g p)) => Eq ((f :@@: g) q p)
-deriving instance Ord (f q (g p)) => Ord ((f :@@: g) q p)
-deriving instance Show (f q (g p)) => Show ((f :@@: g) q p)
-deriving instance Read (f q (g p)) => Read ((f :@@: g) q p)
-
-newtype Par2 (q :: k -> Type) (p :: k)
-    = Par2 { unPar2 :: q p }
-
-deriving instance Eq (q p) => Eq (Par2 q p)
-deriving instance Ord (q p) => Ord (Par2 q p)
-deriving instance Show (q p) => Show (Par2 q p)
-deriving instance Read (q p) => Read (Par2 q p)
-
--------------------------------------------------------------------------------
--- Synonyms for convenience
--------------------------------------------------------------------------------
-
-type D2 = M2 D
-type C2 = M2 C
-type S2 = M2 S
 
 -------------------------------------------------------------------------------
 -- Generic type class
@@ -141,7 +62,24 @@ class Generic (a :: Type) where
     type Rep a = Translate (GHC.Rep a)
 
     to   :: Quote q => Rep a (Code q) x -> Code q a
+    default to :: (Rep a ~ Translate (GHC.Rep a), GHC.Generic a, GGeneric (Rep a), Quote q)
+               => Rep a (Code q) x -> Code q a
+    to = genericTo
+
     from :: Quote q => Code q a -> (Rep a (Code q) x -> Code q r) -> Code q r
+    default from :: (Rep a ~ Translate (GHC.Rep a), GHC.Generic a, GGeneric (Rep a), Quote q)
+      => Code q a -> (Rep a (Code q) x -> Code q r) -> Code q r
+    from = genericFrom
+
+-- | Use GHC generics to implement the 'to' method.
+genericTo :: (Rep a ~ Translate (GHC.Rep a), GHC.Generic a, GGeneric (Rep a), Quote q)
+  => Rep a (Code q) x -> Code q a
+genericTo = ghcGenericTo
+
+-- | Use GHC generics to implement the 'from' method.
+genericFrom :: (Rep a ~ Translate (GHC.Rep a), GHC.Generic a, GGeneric (Rep a), Quote q)
+  => Code q a -> (Rep a (Code q) x -> Code q r) -> Code q r
+genericFrom = ghcGenericFrom
 
 class Generic1 (f :: k -> Type) where
     type Rep1 f :: (Type -> Type) -> k -> Type
@@ -149,28 +87,6 @@ class Generic1 (f :: k -> Type) where
 
     to1   :: Quote q => Rep1 f (Code q) x -> Code q (f x)
     from1 :: Quote q => Code q (f x) -> (Rep1 f (Code q) x -> Code q r) -> Code q r
-
--------------------------------------------------------------------------------
--- Derive our Rep from GHC.Generics.Rep
--------------------------------------------------------------------------------
-
--- | Translate "GHC.Generics" 'GHC.Rep' type into our 'Rep' type.
-type family Translate (f :: k -> Type) :: (Type -> Type) -> k -> Type where
-    Translate (GHC.M1 i c f) = M2 i c (Translate f)
-    Translate (GHC.K1 R c)   = K2 c
-    Translate (f GHC.:+: g)  = Translate f :++: Translate g
-    Translate (f GHC.:*: g)  = Translate f :**: Translate g
-    Translate (GHC.Rec1 f)   = Par2 :@@: f
-    Translate GHC.Par1       = Par2
-    Translate GHC.U1         = U2
-    Translate GHC.V1         = V2
-    Translate (f GHC.:.: g)  = TranslateComp (Par2 :@@: f) g
-    Translate x              = TypeError ('Text "Translate error: " ':<>: 'ShowType x)
-
-type family TranslateComp (f :: (k -> Type) -> k1 -> Type) (g :: k2 -> k1) :: (Type -> Type) -> k -> Type where
-    TranslateComp acc (f GHC.:.: g) = TranslateComp (acc :@@: f) g
-    TranslateComp acc (GHC.Rec1 f)  = acc :@@: f
-    TranslateComp _   x             = TypeError ('Text "Translate :.: error: " ':<>: 'ShowType x)
 
 -------------------------------------------------------------------------------
 -- Example instance(s)
