+++ staged-gg.cabal
@@ -51,10 +51,11 @@ library
     Staged.GHC.Generics.Instances
     Staged.GHC.Generics.TH
     Staged.GHC.Generics.Types
-    Staged.GHC.Generics.GHCish
+    Staged.GHC.Generics.RepTypes
 
   other-modules:
     Generics.Deriving.TH.Internal
     Generics.Deriving.TH.Post4_9
     Staged.GHC.Generics.Internal
     Staged.GHC.Generics.TH.Names
+    Staged.GHC.Generics.GHCish
